+++ src/main/java/com/laba1/ForecastInterface.java
@@ -1,0 +1,14 @@
+package com.laba1;
+
+import java.util.Date;
+import java.util.List;
+
+public interface ForecastInterface {
+    String randomForecast(List<String> forecastsHoroscopeList);
+
+    String randomWeather(Date date, List<String> forecastsWeatherList);
+
+    String randomWeather(List<String> forecastsWeatherList);
+
+    String payBilling(int horoscopeBilling, int weatherBilling);
+}
