+++ src/main/java/com/laba1/ReadFromFileInterface.java
@@ -1,0 +1,10 @@
+package com.laba1;
+
+import java.io.File;
+import java.util.List;
+
+public interface ReadFromFileInterface {
+    File getFileFromResources(String fileName);
+
+    List<String> readTxtFile(String pathname);
+}
