--- src/main/java/laba1/Billing.java
@@ -1,17 +1,0 @@
-package laba1;
-
-import java.io.FileNotFoundException;
-
-public interface Billing {
-  String horoscopeFile = "horoscope.txt";
-  String weatherFile = "weather.txt";
-  InputOutputConsole inputOutputConsole = new InputOutputConsole();
-  InputOutputFile inputOutputFile = new InputOutputFile();
-  Forecast forecast = new Forecast();
-
-  void menu() throws FileNotFoundException;
-
-  void caseHoroscope();
-
-  void caseWeather();
-}
